+++ pages/user/[nameUser]/components/Favorite.jsx
@@ -26,6 +26,10 @@ const Favorite = ({ movie, toast, setArrFavoriteMovie }) => {
     e.preventDefault();
     e.stopPropagation();
     try {
+      if (!user || !accessToken) {
+        toast("Đăng nhập để sử dụng tính năng này");
+        return;
+      }
       const res = await deleteFavoriteMovie(userId, movie._id);
       // console.log(">>> deleteFavoriteMovie <<<", res);
